refactor assessmentform and hooks into subfolder
