+++ clients/assessment_component/src/assessment/pages/AssessmentPage/components/AssessmentForm/hooks/useCreateAssessment.tsx
@@ -1,18 +1,15 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { useParams } from 'react-router-dom'
-import { updateAssessment } from '../../../network/mutations/createOrUpdateAssessment'
-import { CreateOrUpdateAssessmentRequest } from '../../../interfaces/assessment'
+import { createAssessment } from '../../../../../network/mutations/createOrUpdateAssessment'
+import { CreateOrUpdateAssessmentRequest } from '../../../../../interfaces/assessment'
 
-export const useUpdateAssessment = (setError: (error: string | null) => void) => {
+export const useCreateAssessment = (setError: (error: string | null) => void) => {
   const { phaseId } = useParams<{ phaseId: string }>()
   const queryClient = useQueryClient()
 
   return useMutation({
     mutationFn: (assessment: CreateOrUpdateAssessmentRequest) => {
-      return updateAssessment({
-        ...assessment,
-        coursePhaseID: phaseId ?? '',
-      })
+      return createAssessment({ ...assessment, coursePhaseID: phaseId ?? '' })
     },
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['assessments', phaseId] })
