+++ clients/assessment_component/src/assessment/pages/AssessmentPage/components/CategoryAssessment.tsx
@@ -2,7 +2,7 @@ import { useParams } from 'react-router-dom'
 import { useState } from 'react'
 import { ChevronRight, ChevronDown } from 'lucide-react'
 
-import { AssessmentForm } from './AssessmentForm'
+import { AssessmentForm } from './AssessmentForm/AssessmentForm'
 import type { CategoryWithCompetencies } from '../../../interfaces/category'
 import type { Assessment } from '../../../interfaces/assessment'
 import AssessmentStatusBadge from './AssessmentStatusBadge'
